requirements.txt
asgiref==3.5.2
Django==4.0.5
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
importlib-metadata==4.11.4
Markdown==3.3.7
psycopg2-binary==2.9.3
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
zipp==3.8.0
